+++ src/user_preferences.jl
@@ -20,6 +20,7 @@ Check your preferences by calling `get_preferences(key::String)`.
 - `ANTHROPIC_API_KEY`: The API key for the Anthropic API. Get yours from [here](https://www.anthropic.com/).
 - `VOYAGE_API_KEY`: The API key for the Voyage API. Free tier is upto 50M tokens! Get yours from [here](https://dash.voyageai.com/api-keys).
 - `GROQ_API_KEY`: The API key for the Groq API. Free in beta! Get yours from [here](https://console.groq.com/keys).
+- `DEEPSEEK_API_KEY`: The API key for the DeepSeek API. Get \$5 credit when you join. Get yours from [here](https://platform.deepseek.com/api_keys).
 - `MODEL_CHAT`: The default model to use for aigenerate and most ai* calls. See `MODEL_REGISTRY` for a list of available models or define your own.
 - `MODEL_EMBEDDING`: The default model to use for aiembed (embedding documents). See `MODEL_REGISTRY` for a list of available models or define your own.
 - `PROMPT_SCHEMA`: The default prompt schema to use for aigenerate and most ai* calls (if not specified in `MODEL_REGISTRY`). Set as a string, eg, `"OpenAISchema"`.
@@ -46,6 +47,7 @@ Define your `register_model!()` calls in your `startup.jl` file to make them ava
 - `ANTHROPIC_API_KEY`: The API key for the Anthropic API. Get yours from [here](https://www.anthropic.com/).
 - `VOYAGE_API_KEY`: The API key for the Voyage API. Free tier is upto 50M tokens! Get yours from [here](https://dash.voyageai.com/api-keys).
 - `GROQ_API_KEY`: The API key for the Groq API. Free in beta! Get yours from [here](https://console.groq.com/keys).
+- `DEEPSEEK_API_KEY`: The API key for the DeepSeek API. Get \$5 credit when you join. Get yours from [here](https://platform.deepseek.com/api_keys).
 
 Preferences.jl takes priority over ENV variables, so if you set a preference, it will take precedence over the ENV variable.
 
@@ -64,6 +66,7 @@ const ALLOWED_PREFERENCES = ["MISTRALAI_API_KEY",
     "ANTHROPIC_API_KEY",
     "VOYAGE_API_KEY",
     "GROQ_API_KEY",
+    "DEEPSEEK_API_KEY",
     "MODEL_CHAT",
     "MODEL_EMBEDDING",
     "MODEL_ALIASES",
@@ -177,6 +180,10 @@ const VOYAGE_API_KEY::String = @load_preference("VOYAGE_API_KEY",
 
 _temp = get(ENV, "GROQ_API_KEY", "")
 const GROQ_API_KEY::String = @load_preference("GROQ_API_KEY",
+    default=_temp);
+
+_temp = get(ENV, "DEEPSEEK_API_KEY", "")
+const DEEPSEEK_API_KEY::String = @load_preference("DEEPSEEK_API_KEY",
     default=_temp);
 
 _temp = get(ENV, "LOCAL_SERVER", "http://localhost:10897/v1")
@@ -342,6 +349,9 @@ aliases = merge(
         "gllama370" => "llama3-70b-8192",
         "gl70" => "llama3-70b-8192",
         "gmixtral" => "mixtral-8x7b-32768"
+        ## DeepSeek
+        "dschat" => "deepseek-chat",
+        "dscode" => "deepseek-coder"
     ),
     ## Load aliases from preferences as well
     @load_preference("MODEL_ALIASES", default=Dict{String, String}()))
@@ -646,7 +656,17 @@ registry = Dict{String, ModelSpec}(
         GroqOpenAISchema(),
         2.7e-7,
         2.7e-7,
-        "Mistral.ai Mixtral 8x7b, hosted by Groq. Max 32K context. See details [here](https://console.groq.com/docs/models)")
+        "Mistral.ai Mixtral 8x7b, hosted by Groq. Max 32K context. See details [here](https://console.groq.com/docs/models)"),
+    "deepseek-chat" => ModelSpec("deepseek-chat",
+        DeepSeekOpenAISchema(),
+        1.4e-7,
+        2.8e-7,
+        "Deepseek.com-hosted DeepSeekV2 model. Max 32K context. See details [here](https://platform.deepseek.com/docs)"),
+    "deepseek-coder" => ModelSpec("deepseek-coder",
+        DeepSeekOpenAISchema(),
+        1.4e-7,
+        2.8e-7,
+        "Deepseek.com-hosted coding model. Max 16K context. See details [here](https://platform.deepseek.com/docs)")
 )
 
 """